refactor(types): move `SheetOf` and `mapSheet`
