+++ services/scan/src/util/metadata.ts
@@ -1,8 +1,12 @@
 import { encodeHmpbBallotPageMetadata } from '@votingworks/ballot-encoder';
 import { metadataFromBytes } from '@votingworks/ballot-interpreter-vx';
-import { BallotPageMetadata, ElectionDefinition } from '@votingworks/types';
+import {
+  BallotPageMetadata,
+  ElectionDefinition,
+  SheetOf,
+} from '@votingworks/types';
 import { Buffer } from 'buffer';
-import { BallotPageQrcode, SheetOf } from '../types';
+import { BallotPageQrcode } from '../types';
 
 function tryMetadataFromBytes(
   electionDefinition: ElectionDefinition,
