+++ services/scan/test/util/mocks.ts
@@ -4,13 +4,13 @@ import {
   FakeWritable,
   fakeWritable,
 } from '@votingworks/test-utils';
+import { SheetOf } from '@votingworks/types';
 import { throwIllegalValue } from '@votingworks/utils';
 import { ChildProcess } from 'child_process';
 import { EventEmitter } from 'events';
 import { fileSync } from 'tmp';
 import { MaybeMocked, mocked } from 'ts-jest/dist/utils/testing';
 import { BatchControl, BatchScanner } from '../../src/fujitsu_scanner';
-import { SheetOf } from '../../src/types';
 import { writeImageData } from '../../src/util/images';
 import { inlinePool, WorkerOps, WorkerPool } from '../../src/workers/pool';
